% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBins.R
\name{addBins}
\alias{addBins}
\title{addBins}
\usage{
addBins(gr, nbins = 5, bin = NULL)
}
\arguments{
\item{gr}{a gRanges object}

\item{nbins}{a number corresponding to}

\item{bin}{a vector on number}
}
\value{
a gRanges with a new "bin" column
}
\description{
addBins
}
\examples{
addBins(gRanges_test)

# 3 bins instead of 5
addBins(gRanges_test, nbins = 3)

# assign bins using a vector
addBins(gRanges_test, bin = rep(c("a", "b", "c"),
 length.out = length(gRanges_test)))


}
