% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMetricAndArrangeGRanges.R
\name{addMetricAndArrangeGRanges}
\alias{addMetricAndArrangeGRanges}
\title{addMetricAndArrangeGRanges}
\usage{
addMetricAndArrangeGRanges(
  gr,
  myOrder,
  gr_key = "name",
  order_key = "name",
  order_value = "exp",
  shuffle_tie = TRUE
)
}
\arguments{
\item{gr}{an input GRanges}

\item{myOrder}{a data.frame or data.tabe contains the list of genes}

\item{gr_key}{a character ...}

\item{order_key}{a character corresponding to the column
that will be used to sort}

\item{order_value}{a character corresponding to the column
of gene expression}

\item{shuffle_tie}{a boolean Value (TRUE - FALSE),
when TRUE, shuffle the gRanges before sorting}
}
\value{
a GRanges sorted in ascending order
}
\description{
addMetricAndArrangeGRanges
}
\examples{
data("gRanges_test")
ramdomOrder <- data.frame(gene_id = gRanges_test$gene_id,
   value = rnorm(length(gRanges_test)))
addMetricAndArrangeGRanges(gRanges_test,
   ramdomOrder, gr_key = "gene_id",
   order_key = "gene_id", order_value = "value" )


}
