% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_profile.R
\name{average_profile}
\alias{average_profile}
\title{average_profile}
\usage{
average_profile(gr, mycols)
}
\arguments{
\item{gr}{a gRanges input}

\item{mycols}{a character}
}
\value{
a plot
}
\description{
average_profile
}
\examples{
mycols <- paste0("window_", 1:51)
average_profile(gRanges_test, mycols)
}
