% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_profileExp.R
\name{boxplot_profileExp}
\alias{boxplot_profileExp}
\title{boxplot_profileExp}
\usage{
boxplot_profileExp(x, trans_func = function(x) log10(x + 1))
}
\arguments{
\item{x}{a gRanges input}

\item{trans_func}{a function which applicate log10 transformation}
}
\value{
a boxplot
}
\description{
boxplot_profileExp
}
\examples{
boxplot_profileExp(gRanges_test$exp)
}
