% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_plot.R
\name{global_plot}
\alias{global_plot}
\title{global_plot}
\usage{
global_plot(x, mycols = paste0("window_", 1:51))
}
\arguments{
\item{x}{a gRanges input}

\item{mycols}{a character}
}
\value{
a png
}
\description{
global_plot
}
\examples{
bw <- "G11+_fed_R6.bigWig"
mysample <- list_file$id_bed[which(list_file$id_bw == bw)]
global_plot(gRanges_test)
}
