% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegendGeneType.R
\name{legned_gene_type}
\alias{legned_gene_type}
\title{legned_gene_type}
\usage{
legned_gene_type(gr)
}
\arguments{
\item{gr}{a gRanges object}
}
\value{
a legend for plot of gene_type
}
\description{
legned_gene_type
}
\examples{
legned_gene_type(my_tss)
}
