% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_file.R
\docType{data}
\name{list_file}
\alias{list_file}
\title{Sus scrofa chr18 expression data combining a "bigwig" file and the "bed"
annotation file}
\format{
A \code{csv file} of the 31 rows of samples
and id corresponding for each sample
}
\source{
This this test set was generated from the metadata,
The csv file contains 2 columns:
the bigwig identifier and its equivalent in the starting metadata
}
\usage{
list_file
}
\description{
Sus scrofa chr18 expression data combining a "bigwig" file and the "bed"
annotation file
}
\keyword{datasets}
