% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanColors.R
\name{mean_color}
\alias{mean_color}
\title{mean_color}
\usage{
mean_color(colors)
}
\arguments{
\item{colors}{vector of colors}
}
\value{
a color
}
\description{
mean_color
}
\examples{
mean_color(c("#000000FF", "#FFFFFF00", "#FFFF00FF", "#FF0000FF"))

# works with color names
mean_color(c("blue", "red"))

# Mix color names and HTML codes
mean_color(c("blue", "red", "#FFFF00FF"))

# works without alpha channel in inputs (but outputs an alpha channel):
mean_color(c("#000000", "#FFFFFF", "#FFFF00", "#FF0000"))
}
