% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanColor.R
\name{meanColor}
\alias{meanColor}
\title{meanColor}
\usage{
meanColor(colors)
}
\arguments{
\item{colors}{a vector of colors}
}
\value{
a single color value
}
\description{
Return the average color of a vector of colors, computed in the RGB space.
}
\details{
Input colors can be either in html or color name formats.
The alpha channel is supported but optional.
}
\examples{
meanColor(c("#000000FF", "#FFFFFF00", "#FFFF00FF", "#FF0000FF"))

# works with color names
meanColor(c("blue", "red"))

# Mix color names and HTML codes
meanColor(c("blue", "red", "#FFFF00FF"))

# works without alpha channel in inputs (but outputs an alpha channel):
meanColor(c("#000000", "#FFFFFF", "#FFFF00", "#FF0000"))
}
\seealso{
\code{\link[epistack]{redimMatrix}}
}
