% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAverageProfile.R
\name{plotAverageProfile}
\alias{plotAverageProfile}
\title{plotAverageProfile()}
\usage{
plotAverageProfile(
  rse,
  assay = NULL,
  x_labels = c("Before", "Anchor", "After"),
  palette = colorRampPalette(c("magenta", "black", "green")),
  alpha_for_se = 0.25,
  error_type = c("sd", "sem"),
  reversed_z_order = FALSE,
  ylim = NULL,
  pattern = NULL
)
}
\arguments{
\item{rse}{a RangedSummarizedExperiment input. Aletrnatively: can be a
GRanges object
(for backward compatibility, \code{pattern} will be required).}

\item{assay}{specify the name of the assay to plot,
that should match one of \code{assayNames(rse)}.}

\item{x_labels}{x-axis labels.}

\item{palette}{a color palette function,
by default: \code{colorRampPalette(c("magenta", "black", "green"))}}

\item{alpha_for_se}{the transparency (alpha) value for
the error band.}

\item{error_type, }{can be either \code{sd} (standard deviation)
or \code{sem} (standard error of the mean). Default: \code{sem}.}

\item{reversed_z_order}{should the z-order of the curves be reversed
(i.e. first or last bin on top?)}

\item{ylim}{a vector of two numbers corresponding
to the y-limits of the plot}

\item{pattern}{only if \code{rse} is of class GRanges.
A single character that should match
metadata of \code{rse} (can be a regular expression).}
}
\value{
Display a plot.
}
\description{
Plot the average stack profiles +/-  error (sd or sem).
If a \code{bin} column is present in \code{rowRanges(rse)},
one average profile is drawn for each bin.
}
\examples{
data("stackepi")
plotAverageProfile(stackepi)

}
