% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinning.R
\name{plotBinning}
\alias{plotBinning}
\title{plotBinning}
\usage{
plotBinning(
  gr,
  target_height = 650,
  palette = colorRampPalette(c("magenta", "black", "green"))
)
}
\arguments{
\item{gr}{a gRanges object containing a "bin" column}

\item{target_height}{an integer, the approximate number of pixels of the final plot.
Used to reduce the matrix before plotting to improve performance and reduce overploting artefacts.}

\item{palette}{a function taking a number has a first argument, and returning a vector of colors.}
}
\value{
a plot
}
\description{
plotBinning
}
\examples{
gr <- gRanges_test
gr <- addBins(gr, nbins = 3)
plot_bin <- plotBinning(gr)


gr2 <- data.frame(bin = rep(c(1,2,3,4), each = 5))
plotBinning(gr2, palette = colorRampPalette(c("blue4", "forestgreen", "coral3", "goldenrod")))



}
