% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxMetric.R
\name{plotBoxMetric}
\alias{plotBoxMetric}
\title{plotBoxMetric}
\usage{
plotBoxMetric(
  gr,
  metric = "expr",
  title = "Metric",
  trans_func = function(x) x,
  ylim = NULL,
  ylab = "metric",
  palette = colorRampPalette(c("magenta", "black", "green"))
)
}
\arguments{
\item{gr}{a GRanges input}

\item{trans_func}{a function to transform value of x.
Useful to apply log10 transformation (i.e. with \code{trans_func = function(x) log10(x+1)}).}

\item{colorPalette}{a function of parameter \code{n} that returns a palette of \code{n} colors.}
}
\value{
a boxplot
}
\description{
plotBoxMetric
}
\examples{
boxplot_profileExp(gRanges_test$exp)

}
