% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEpistack.R
\name{plotEpistack}
\alias{plotEpistack}
\title{plotEpistack}
\usage{
plotEpistack(
  gr,
  patterns = "^window_",
  tints = "gray",
  titles = "window",
  x_labels = c("-2.5kb", "TSS", "+2.5kb"),
  zlim = c(0, 1),
  ylim = NULL,
  metric_col = "expr",
  metric_title = "Metric",
  metric_label = "metric",
  metric_transfunc = function(x) x,
  bin_palette = colorRampPalette(c("magenta", "black", "green")),
  npix_height = 650,
  n_core = 1,
  high_mar = c(2.5, 0.6, 4, 0.6),
  low_mar = c(2.5, 0.6, 0.3, 0.6)
)
}
\arguments{
\item{gr}{a gRanges input}

\item{patterns}{a character vector of patterns. Used to subset the \code{gr} for each individual plots when
multiple tracks are presents (i.e. a ChIP-seq and its Input control).}

\item{tints}{a colour vector of length \code{patterns}, to tint each epistack.}
}
\value{
a png
}
\description{
plotEpistack
}
\examples{
bw <- "G11+_fed_R6.bigWig"
mysample <- list_file$id_bed[which(list_file$id_bw == bw)]
global_plot(gRanges_test)
}
