% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpressionProfile.R
\name{plotExpressionProfile}
\alias{plotExpressionProfile}
\title{plotExpressionProfile}
\usage{
plotExpressionProfile(x, trans_func = function(x) log10(x + 1), title = "")
}
\arguments{
\item{x}{a GRanges input}

\item{trans_func}{a function which applicate log10 transformation}

\item{title}{a character corresponding to the title of the plot}
}
\value{
a plot
}
\description{
plotExpressionProfile
}
\examples{
plotExpressionProfile(gRanges_test$exp, title = "Expression")
}
