% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegendGeneType.R
\name{plotGeneTypeLegend}
\alias{plotGeneTypeLegend}
\title{plotGeneTypeLegend}
\usage{
plotGeneTypeLegend(gr, palette = grDevices::rainbow)
}
\arguments{
\item{gr}{a gRanges object}

\item{palette}{a function taking a number has a first argument, and returning a vector of colors.}
}
\value{
a legend for plot of gene_type
}
\description{
plotGeneTypeLegend
}
\examples{
library(GenomicRanges)
gr <- GRanges(
    rep("chr1", 10),
    IRanges(seq(1, 901, by = 100), seq(100, 1000, by = 100)),
    gene_type = c(rep("protein_coding", 5), rep("linc_RNA", 5))
)
plotGeneTypeLegend(gr)
}
