% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetric.R
\name{plotMetric}
\alias{plotMetric}
\title{plotMetric}
\usage{
plotMetric(
  x,
  trans_func = function(x) x,
  title = "Metric",
  ylim = NULL,
  xlab = NULL
)
}
\arguments{
\item{x}{a GRanges input}

\item{trans_func}{a function to transform value of x.
Useful to apply log10 transformation (i.e. with \code{trans_func = function(x) log10(x+1)}).}

\item{title}{a character corresponding to the title of the plot}
}
\value{
a plot
}
\description{
plotMetric
}
\examples{
plotExpressionProfile(gRanges_test$exp, title = "Expression")
}
