% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim_matrix.R
\name{redim_matrix}
\alias{redim_matrix}
\title{redim_matrix}
\usage{
redim_matrix(
  mat,
  target_height = 100,
  target_width = 100,
  summary_func = function(x) mean(x, na.rm = TRUE),
  output_type = 0,
  n_core = 1
)
}
\arguments{
\item{mat}{a matrix: the input matrix}

\item{target_height}{a number: the length of the output matrix}

\item{target_width}{a number: the width of the output matrix}

\item{summary_func}{a function: the mean function with "na.rm = TRUE"}

\item{output_type}{a number: vapply style}

\item{n_core}{a number: parallel processing}
}
\value{
a matrix: a resized matrix
}
\description{
this function resizes the input matrix by applying the rounded average
on the columns
}
\examples{
data("gRanges_test")
mat <- S4Vectors::mcols(gRanges_test)
whichCols <- grepl("^window_", colnames(mat))
mat <- as.matrix(mat[, whichCols])
dim(mat)
smallMat <- redim_matrix(mat, target_height = 10, target_width = ncol(mat))
dim(smallMat)

mat <- matrix(sample(1:40,100,replace=TRUE),nrow=10,ncol=10)
dim(mat)
smallMat <- redim_matrix(mat, target_height = 5, target_width = ncol(mat),
   summary_func = function(x) max(x, na.rm = TRUE))
dim(smallMat)
}
