% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackepi.R
\docType{data}
\name{stackepi}
\alias{stackepi}
\title{epistack example and test dataset}
\format{
A \code{GRanges} of the 693 rows and 54 metadata columns
}
\source{
This dataset was generated from ENSEMBL annotation data
and data generated by our lab (publicly available soon).
}
\usage{
data("stackepi")
}
\description{
DNA methylation profiles (from MBD-seq data) arround transcription
start sites of the 693 chr18 genes annotated on the pig genome
(Sscrofa11.1), as well as gene expression levels in
Transcript Per Million (TPM) measured by RNA-seq in the same
duodenum sample.
}
\keyword{datasets}
