% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBins.R
\name{addBins}
\alias{addBins}
\title{addBins}
\usage{
addBins(gr, nbins = 5, bin = NULL)
}
\arguments{
\item{gr}{a gRanges object}

\item{nbins}{a number corresponding to}

\item{bin}{a vector on number}
}
\value{
a gRanges with a new "bin" column
}
\description{
addBins
}
\examples{
data <- transform(data, log2fc_line = as.numeric(log2fc_line))
mat <- data[data$adjpval_line <= 0.01, c("gene", "log2fc_line", "gene_type")]
TSS <- promoters(TSS, upstream = 0, downstream = 1)
my_tss <- addMetricAndArrangeGRanges(TSS, mat, order_key = "gene", order_value = "log2fc_line")
my_tss <- addBins(my_tss)
my_tss
}
