% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBins.R
\name{addBins}
\alias{addBins}
\title{addBins()}
\usage{
addBins(gr, nbins = 5L, bin = NULL)
}
\arguments{
\item{gr}{a GRanges object.}

\item{nbins}{an integer number, the number of bins.}

\item{bin}{a vector containing pre-determined bins, in the same
order as \code{gr}.}
}
\value{
the \code{gr} GRanges object with a new \code{bin} metadata column
}
\description{
Add an optional bin metadata column to \code{gr}, to serve as annotations
for the epistack plots.
}
\details{
\code{nbins} is taken into account only if \code{bin} is \code{NULL}.
\code{gr} should be sorted first, usually with the
\code{addMetricAndArrangeGRanges()}
function. \code{addBin(gr, bin = vec)} is equivalent to
\code{gr$bin <- vec}, while
\code{addBin(gr, nbins = 5)} will create 5 bins of equal size based on
\code{gr}
order.
}
\examples{
data("gRanges_test")
addBins(gRanges_test)

# 3 bins instead of 5
addBins(gRanges_test, nbins = 3)

# assign bins using a vector
addBins(gRanges_test, bin = rep(c("a", "b", "c"),
 length.out = length(gRanges_test)))
}
\seealso{
\code{\link[epistack]{addMetricAndArrangeGRanges}}
}
