% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRanges_test.R
\docType{data}
\name{gRanges_test}
\alias{gRanges_test}
\title{Sus scrofa chr18 expression data combining a "bigwig" file and the "bed"
annotation file}
\format{
A \code{gRanges} of the 693 rows of chromosome 18:
}
\source{
This dataset was generated from ENSEMBL annotation data and
a bigwig file generated by nfcore/chipseq
}
\usage{
data("gRanges_test")
}
\value{
A \code{matrix}.
}
\description{
Sus scrofa chr18 expression data combining a "bigwig" file and the "bed"
annotation file
}
\keyword{datasets}
