% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_file.R
\docType{data}
\name{list_file}
\alias{list_file}
\title{Sus scrofa chr18 expression data combining a "bigwig" file and the "bed"
annotation file}
\format{
A \code{data.frame} of 24 samples + 4 Tem-methyl and 4 Tem-input
with bigwig files after the BayMeth transformation (methylation ratio)
}
\source{
This dataset was generated from starting data and
a bigwig file generated by BayMeth method
}
\usage{
data("list_file")
}
\value{
A \code{data.frame}.
}
\description{
Sus scrofa chr18 expression data combining a "bigwig" file and the "bed"
annotation file
}
\keyword{datasets}
