% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAverageProfile.R
\name{plotAverageProfile}
\alias{plotAverageProfile}
\title{plotAverageProfile()}
\usage{
plotAverageProfile(
  gr,
  what_pattern = "^window_",
  xlabels = c("-2.5kb", "TSS", "+2.5kb"),
  palette = colorRampPalette(c("magenta", "black", "green")),
  alpha_for_se = 0.25,
  reversed_z_order = FALSE,
  ylim = NULL
)
}
\arguments{
\item{gr}{a GRanges input}

\item{what_pattern}{a single character that should match
metadata of \code{gr} (can be a regular expression).}

\item{xlabels}{x-axis labels.}

\item{palette}{a color palette function,
by default: \code{colorRampPalette(c("magenta", "black", "green"))}}

\item{alpha_for_se}{the transparency (alpha) value for
standard error of the mean band.}

\item{reversed_z_order}{should the z-order of the curves be reversed
(i.e. first or last bin on top?)}

\item{ylim}{a vector of two numbers corresponding
to the y-limits of the plot}
}
\description{
Plot the average stack profiles +/- standard error of the
means. If a \code{bin} column is present in \code{gr}, one average profile
is drawn for each bin.
}
\examples{
data("stackepi")
plotAverageProfile(stackepi)

}
