% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAverageProfile.R
\name{plotAverageProfile}
\alias{plotAverageProfile}
\title{plotAverageProfile}
\usage{
plotAverageProfile(
  gr,
  what_pattern = "^window_",
  xlabels = c("-2.5kb", "TSS", "+2.5kb"),
  colorPalette = colorRampPalette(c("magenta", "black", "green")),
  alphaForSe = 0.25,
  reversedZOrder = FALSE,
  ylim = NULL
)
}
\arguments{
\item{gr}{a gRanges input}

\item{ylim}{a vector of number corresponding to the limits of the values to be displayed}

\item{mycols}{a character}

\item{labels}{a vector of character corresponding to the labels of the x-axis}
}
\value{
a plot
}
\description{
plotAverageProfile
}
\examples{
mycols <- paste0("window_", 1:51 )
plotAverageProfile(gRanges_test, mycols, ylim = c(0, 2))



}
