% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAverageProfile.R
\name{plotAverageProfile}
\alias{plotAverageProfile}
\title{plotAverageProfile}
\usage{
plotAverageProfile(
  gr,
  what_pattern = "^window_",
  xlabels = c("-2.5kb", "TSS", "+2.5kb"),
  colorPalette = colorRampPalette(c("magenta", "black", "green")),
  alphaForSe = 0.25,
  reversedZOrder = FALSE,
  ylim = NULL
)
}
\arguments{
\item{gr}{a gRanges input}

\item{what_pattern}{a character vector of column prefixes
(can be regular expressions)}

\item{xlabels}{x axis labels}

\item{colorPalette}{color palette,
by default: colorRampPalette(c("magenta", "black", "green"))}

\item{alphaForSe}{transparency value for standard error of the mean bend}

\item{reversedZOrder}{should the z-order of the curves be reversed}

\item{ylim}{a vector of number corresponding
to the limits of the values to be displayed}
}
\value{
a png
}
\description{
plotAverageProfile
}
\examples{
data("gRanges_test")
plotAverageProfile(gRanges_test)

}
