% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinning.R
\name{plotBinning}
\alias{plotBinning}
\title{plotBinning()}
\usage{
plotBinning(
  rse,
  target_height = 650,
  palette = colorRampPalette(c("magenta", "black", "green"))
)
}
\arguments{
\item{rse}{a RangedSummarizedExperiment input with a column \code{bin} in
\code{rowRanges(rse)}.
Alternatively (for backward compatibility),
a GRanges object or any object such as \code{rse$bin} exists.}

\item{target_height}{an integer, the approximate height (in pixels)
of the final plot. Used to avoid overplotting artefacts.}

\item{palette}{a function taking a number as a first argument,
and returning a vector of colors.}
}
\value{
Display a plot.
}
\description{
Plot a vertical color bar of the \code{bin} column.
}
\examples{
data("stackepi")
gr <- stackepi
gr <- addBins(gr, nbins = 3)
plot_bin <- plotBinning(gr)

gr2 <- data.frame(bin = rep(c(1,2,3,4), each = 5))
plotBinning(gr2, palette = colorRampPalette(c("blue4", "forestgreen", "coral3", "goldenrod")))

}
