% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxMetric.R
\name{plotBoxMetric}
\alias{plotBoxMetric}
\title{plotBoxMetric}
\usage{
plotBoxMetric(
  gr,
  metric = "expr",
  title = "Metric",
  trans_func = function(x) x,
  ylim = NULL,
  ylab = "metric",
  palette = colorRampPalette(c("magenta", "black", "green"))
)
}
\arguments{
\item{gr}{a GRanges input}

\item{metric}{name of the column containing epigenetic profile score}

\item{title}{a character corresponding to metric title}

\item{trans_func}{a function to transform value of x
Useful to apply log10 transformation (i.e. with \code{trans_func = function(x) log10(x+1)}).}

\item{ylim}{limit of the y axis; format: ylim = c (min, max)}

\item{ylab}{y axis title}

\item{palette}{a function of parameter \code{n} that returns a palette of \code{n} colors.}
}
\value{
a boxplot
}
\description{
plotBoxMetric
}
\examples{
plotBoxMetric(
       gRanges_test,
       trans_func = function(x) x,
       metric = "exp",
       title = "Metric"
  )


}
