% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEpistack.R
\name{plotEpistack}
\alias{plotEpistack}
\title{plotEpistack}
\usage{
plotEpistack(
  gr,
  patterns = "^window_",
  tints = "gray",
  titles = "window",
  x_labels = c("-2.5kb", "TSS", "+2.5kb"),
  zlim = c(0, 1),
  ylim = NULL,
  metric_col = "expr",
  metric_title = "Metric",
  metric_label = "metric",
  metric_transfunc = function(x) x,
  bin_palette = colorRampPalette(c("magenta", "black", "green")),
  npix_height = 650,
  n_core = 1,
  high_mar = c(2.5, 0.6, 4, 0.6),
  low_mar = c(2.5, 0.6, 0.3, 0.6)
)
}
\arguments{
\item{gr}{a granges input}

\item{tints}{color}

\item{x_labels}{x axis labels}

\item{zlim}{the minimum and maximum z values for which colors should be plotted}

\item{ylim}{limit of the y axis; format: ylim = c (min, max)}

\item{metric_col}{a character corresponding to epigenetic profile score}

\item{metric_title}{name of the score column}

\item{metric_label}{label of the score column}

\item{metric_transfunc}{a function to transform value of x
Useful to apply log10 transformation (i.e. with \code{trans_func = function(x) log10(x+1)}).}

\item{bin_palette}{a palette of color}

\item{n_core}{a number : parallel processing}

\item{high_mar}{a vector of numerical values corresponding to the margins of the top figures}

\item{low_mar}{a vector of numerical values corresponding to the margins of the bottom figures}
}
\value{
a png
}
\description{
plotEpistack
}
\examples{
plotEpistack(gRanges_test, metric_col = "exp", metric_transfunc = function(x) log10(x+1))

}
