% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneType.R
\name{plotGeneType}
\alias{plotGeneType}
\title{plotGeneType}
\usage{
plotGeneType(gr, palette = grDevices::rainbow)
}
\arguments{
\item{gr}{a gRanges object containing a "gene_type" column}

\item{palette}{a function taking a number has a first argument, and returning a vector of colors.}
}
\value{
a plot
}
\description{
plotGeneType
}
\examples{
library(GenomicRanges)
gr <- GRanges(
    rep("chr1", 10),
    IRanges(seq(1, 901, by = 100), seq(100, 1000, by = 100)),
    gene_type = c(rep("protein_coding", 5), rep("linc_RNA", 5))
)
plotGeneType(gr)
}
