% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneType.R
\name{plotGeneType}
\alias{plotGeneType}
\title{plot_gene_type}
\usage{
plotGeneType(gr)
}
\arguments{
\item{gr}{a gRanges object containing a "gene_type" column}
}
\value{
a plot
}
\description{
plot_gene_type
}
\examples{
plot_gene_type(my_tss)

}
