% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneType.R
\name{plotGeneType}
\alias{plotGeneType}
\title{plotGeneType}
\usage{
plotGeneType(gr, palette = RColorBrewer::brewer.pal)
}
\arguments{
\item{gr}{a gRanges object containing a "gene_type" column}

\item{palette}{a function taking a number has a first argument, and returning a vector of colors.}
}
\value{
a plot
}
\description{
plotGeneType
}
\examples{
#plotGeneType(my_tss)

}
