% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegendGeneType.R
\name{plotGeneTypeLegend}
\alias{plotGeneTypeLegend}
\title{plotGeneTypeLegend}
\usage{
plotGeneTypeLegend(gr, palette = RColorBrewer::brewer.pal)
}
\arguments{
\item{gr}{a gRanges object}

\item{palette}{a function taking a number has a first argument, and returning a vector of colors.}
}
\value{
a legend for plot of gene_type
}
\description{
plotGeneTypeLegend
}
\examples{
#plotGeneTypeLegend(gRanges_test)
}
