% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegendGeneType.R
\name{plotGeneTypeLegend}
\alias{plotGeneTypeLegend}
\title{plotGeneTypeLegend}
\usage{
plotGeneTypeLegend(gr)
}
\arguments{
\item{gr}{a gRanges object}
}
\value{
a legend for plot of gene_type
}
\description{
plotGeneTypeLegend
}
\examples{
#plotGeneTypeLegend(gRanges_test)
}
