% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedProfile.R
\name{plotStackedProfile}
\alias{plotStackedProfile}
\title{plotStackedProfile}
\usage{
plotStackedProfile(
  gr,
  what_pattern = "^window_",
  x_labels = c("-2.5kb", "TSS", "+2.5kb"),
  title = "",
  min_value = NULL,
  max_value = NULL,
  palette = function(n) viridisLite::viridis(n, direction = -1),
  target_height = 650,
  summary_func = function(x) mean(x, na.rm = TRUE),
  n_core = 1
)
}
\arguments{
\item{gr}{a GRanges input}

\item{what_pattern}{a character}

\item{x_labels}{a character vectors with column labels to use}

\item{title}{a character corresponding to the title of the heatmap}

\item{min_value}{a number}

\item{max_value}{a number}

\item{palette}{function to determine the palette of heatmap color}

\item{target_height}{a number}

\item{summary_func}{function whitch applicate
the mean function to the gRanges input}

\item{n_core}{a number : parallel processing}
}
\value{
an heatmap
}
\description{
plotStackedProfile
}
\examples{
plotStackedProfile(gRanges_test,
   max_value = 1,
   target_height = length(gRanges_test),
   title = "DNA methylation")
}
