% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackProfileLengend.R
\name{plotStackProfileLegend}
\alias{plotStackProfileLegend}
\title{plotStackProfileLegend()}
\usage{
plotStackProfileLegend(
  zlim,
  palette = colorRampPalette(c("white", "grey", "black")),
  title = NA
)
}
\arguments{
\item{zlim}{the limits of the values to be displayed.
Format: \code{c(min, max)}}

\item{palette}{a palette of color,
(i.e. a function of parameter n that should retrun n colors).}

\item{title}{an optionnal title to be display bellow the color legend.}
}
\description{
Utility function to plot the corresponding legend key of
\code{plotStackProfile()}'s plots.
}
\examples{
plotStackProfileLegend(zlim = c(0, 2),
    palette = colorRampPalette(c("white", "grey", "black")))

}
\seealso{
\code{\link[epistack]{plotStackProfile}}
}
