% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedProfileLengend.R
\name{plotStackProfileLegend}
\alias{plotStackProfileLegend}
\title{plotStackProfileLegend}
\usage{
plotStackProfileLegend(
  zlim,
  palette = colorRampPalette(c("white", "grey", "black"))
)
}
\arguments{
\item{zlim}{the limits of the values to be displayed}

\item{palette}{color palette,
by default: colorRampPalette(c ("white", "gray", "black"))}
}
\value{
a png
}
\description{
plotStackProfileLegend
}
\examples{
plotStackProfileLegend(zlim = c(0, 2), palette = colorRampPalette(c("white", "grey", "black")))


}
