% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackepi.R
\docType{data}
\name{stackepi}
\alias{stackepi}
\title{epistack example and test dataset}
\format{
For \code{stackepi}: A \code{RangedSummarizedExperiment} of the 693 rows,
2 rows metadata columns, and one assay containing the DNA methylation
signal.
For \code{stackepi_gr}: A \code{GRanges} of the 693 rows and 54 metadata
columns, kept for unit-testing backward-compatibility.
}
\source{
This dataset was generated from ENSEMBL annotation data
and data generated by our lab (publicly available soon).
}
\usage{
data("stackepi")
data("stackepi_gr")
}
\description{
DNA methylation profiles (from MBD-seq data) arround transcription
start sites of the 693 chr18 genes annotated on the pig genome
(Sscrofa11.1), as well as gene expression levels in
Transcript Per Million (TPM) measured by RNA-seq in the same
duodenum sample.
}
\seealso{
\code{\link[epistack]{epistack::GRanges2RSE}}
}
\keyword{datasets}
